python/hydra_phase0_tpu/pyproject.toml
[project]
name = "hydra-phase0-tpu"
version = "0.1.0"
description = "Hydra Phase 0 TPU behavioral cloning trainer"
requires-python = ">=3.11"
dependencies = [
  "jax>=0.9.1",
  "flax>=0.12.5",
  "optax>=0.2.6",
  "orbax-checkpoint>=0.11.33",
  "numpy>=1.26",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "PyYAML>=6.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
